Add more xpaths to find dataset URLs in ECHO feed.
Currently in production, only 25 of 2327 datasets in the ADE have a link
in their title. This change adds links for more than 1800 results by
adding more xpaths where a URL can be taken from.

Many datasets still do not have any links; from a random sampling of a
few of them, the ECHO feed itself does not have URLs for these
datasets. In order to provide the users with *something* useful in these
cases, the default value for ECHO URLs is now the ECHO home page, so
every ECHO dataset will be rendered with at least *some* link in the ADE
results.

Additional minor changes:

- wrap the block comment at the beginning of the column to 80 characters

- remove an unecessary '/' from the second xpath

- use single quotes for the xpaths strings, to make more consistent with
  the rest of the file

[Pivotal 77676782]