Remove ¿ characters and only re-encode if invalid
For some datasets, the feed we harvest contains characters invalid for
UTF-8. This is fixed by the call to `String#encode`; see [here][0] for
more detail.

However, calling `encode` for every string value changes the text even
in cases where the source was perfectly valid UTF-8, due to the second
parameter to `encode` (where we pass in `'binary'`). This leads to
certain valid UTF-8 characters being stripped out (such as é) because
they are represented by multiple bytes, but when we attempt to convert
it *from* binary, they cannot be treated as multi-byte characters.

The datasets where this is important for the [user story][1], datasets
from PDC, are valid UTF-8, but the feed contains some ¿ characters,
which is valid, but looks wrong in the search interface.

Therefore, the solution is to call `encode` iff the string is invalid,
and then remove ¿ characters we don't want showing up in the search
results. This preserves the behavior of cleaning up feeds with invalid
characters, and fixes the behavior of removing acceptable characters.

[0]: http://robots.thoughtbot.com/fight-back-utf-8-invalid-byte-sequences
[1]: https://www.pivotaltracker.com/story/show/90440782