Refactor spatial_coverage_invalid? to valid_solr_spatial_coverage?
Solr fails to accept a document if the `spatial_coverages` describes a
"bounding box" where the north and south values are both 90 or -90,
i.e., it is really a point on the north or south pole, but the east and
west values do not match each other. If the north and south are both 90,
and the east and west values match, Solr can accept the document, just
as it can accept any point value.

To experiment with spatial_coverages that solr can accept, the following
set of commands can be run on a machine with Solr listening on port
8983 (change the compass point values as appropriate for your testing):

export NORTH=0; export SOUTH=0; export EAST=0; export WEST=0; curl "http://localhost:8983/solr/nsidc_oai/update?commit=true" -H "Content-Type: text/xml; charset=utf-8" --data "<add>
  <doc>
    <field name='authoritative_id'>12234abcd</field>
    <field name='spatial_coverages'>$SOUTH $WEST $NORTH $EAST</field>
  </doc>
</add>"