Add the error class to the error output
I concluded that the error in the PDC harvest job was `Errno::ETIMEDOUT`
based on searching the internet for the error message, 'Connection timed
out - connect(2) for "www.polardata.ca" port 80'.

If that is not really the exception we are looking for, adding the error
class should help debugging in the future.