Add harvester for DataONE.org
Include spatial restrictions in the URL because there are over 750k
records at dataone.org, and over 28k with at least a partial bounding
box over 45 degrees North.

Multiple selectors use '.' as the xpath because the records in the Data
ONE feed do not have time or geospatial elements with children; rather,
values like "endDate" and "northBoundCoord" are all direct children of
the "doc" element, i.e., they are at the root of the record.

https://www.pivotaltracker.com/story/show/77763710