Updated paleo ingest defaults/values per PO request
Update default time value to blank/undefined as PO does not want the current date to be the default for this data center.

Updated the displayed data center name from NCDC Paleo to NOAA Paleo.  PO is going to double check the name they'd like us to use - avoid refactoring the internal reference name for now as it may change again.