Iterate over test definitions, instead of expect
A test will fail the first time an `expect` fails; when calling `expect`
many times in one test definition, there could be multiple failures, but
only the first one will be reported. Additionally, by looping over all
the key/value pairs and including the key in the `it` string, it is more
clear from the rspec output which assertions are failing.