Fix strange temporal duration facet values
For a few datasets, `date_range` is returning `nil` here. This leads to
an error in calculating the duration for faceting purposes, and the
string describing the duration (and ending up in the interface) is the
original coverage string from the feed instead of "Not specified".

By converting the result of `date_range` to a string before calling
`empty?`, `nil` is return from `get_temporal_duration`, and the facet
duration string ends up as "Not specified".