`rake default` - always run rubocop *and* tests
Travis-CI runs the default rake task. In the previous configuration, if
rubocop encountered an error, the unit tests were not run. This hid the
fact that unit tests were actually broken in [build 150][0]. We want to
always run both rubocop and unit tests, so that all build issues are
apparent.

By running each task in its own process, rather than as prerequisites of
the `default` task, we ensure that each task is run, and can choose to
exit a failing status if any of the individual tasks failed (taken from
the [The Rake Field Manual][1]).

[0]: https://travis-ci.org/nsidc/search-solr-tools/builds/90809018
[1]: http://www.rakefieldmanual.com/fascicles/002-ignore-failed-tasks.html